refactor: introduce cluster scheduler hierarchy
