+++ src/queens/schedulers/__init__.py
@@ -21,6 +21,7 @@ from typing import TYPE_CHECKING
 from queens.utils.imports import extract_type_checking_imports, import_class_from_class_module_map
 
 if TYPE_CHECKING:
+    from queens.schedulers._cluster_base import _BaseCluster
     from queens.schedulers._scheduler import Scheduler
     from queens.schedulers.cluster import Cluster
     from queens.schedulers.cluster_local import ClusterLocal
